Move download URLs and views to datasets/
This moves all data download URLs to datasets/ so that they use the
same prefix as other dataset pages.

This leaves only redirect code in catalogue so the app can be removed
entirely once we no longer need to support old URLs.
