+++ dataworkspace/dataworkspace/apps/datasets/models.py
@@ -354,7 +354,7 @@ class SourceView(BaseSource):
 
     def get_absolute_url(self):
         return reverse(
-            'catalogue:dataset_source_view_download',
+            'datasets:dataset_source_view_download',
             args=(self.dataset.id, self.id),
         )
 
@@ -424,7 +424,7 @@ class SourceLink(TimeStampedModel):
 
     def get_absolute_url(self):
         return reverse(
-            'catalogue:dataset_source_link_download',
+            'datasets:dataset_source_link_download',
             args=(self.dataset.id, self.id),
         )
 
@@ -457,7 +457,7 @@ class CustomDatasetQuery(TimeStampedModel):
 
     def get_absolute_url(self):
         return reverse(
-            'catalogue:dataset_query_download',
+            'datasets:dataset_query_download',
             args=(self.dataset.id, self.id),
         )
