+++ dataworkspace/dataworkspace/apps/datasets/urls.py
@@ -11,6 +11,26 @@ urlpatterns = [
         name='dataset_detail',
     ),
     path(
+        '<str:dataset_uuid>/link/<str:source_link_id>/download',
+        login_required(views.SourceLinkDownloadView.as_view()),
+        name='dataset_source_link_download',
+    ),
+    path(
+        '<str:dataset_uuid>/view/<str:source_id>/download',
+        login_required(views.SourceViewDownloadView.as_view()),
+        name='dataset_source_view_download',
+    ),
+    path(
+        '<str:dataset_uuid>/query/<int:query_id>/download',
+        login_required(views.CustomDatasetQueryDownloadView.as_view()),
+        name='dataset_query_download',
+    ),
+    path(
+        '<str:dataset_uuid>/reference/<str:format>/download',
+        login_required(views.ReferenceDatasetDownloadView.as_view()),
+        name='reference_dataset_download',
+    ),
+    path(
         '<str:dataset_uuid>/eligibility-criteria',
         login_required(views.eligibility_criteria_view),
         name='eligibility_criteria',
