+++ dataworkspace/dataworkspace/templates/datasets/referencedataset_detail.html
@@ -52,7 +52,7 @@
         <tbody>
           <tr class="govuk-table__row">
             <td class="govuk-table__cell">
-              <a class="govuk-link" href="{% url 'catalogue:reference_dataset_download' object.uuid 'json'%}">
+              <a class="govuk-link" href="{% url 'datasets:reference_dataset_download' object.uuid 'json'%}">
                 {{ object.slug }}-{{ object.published_version }}.json
               </a>
             </td>
@@ -62,7 +62,7 @@
           </tr>
           <tr class="govuk-table__row">
             <td class="govuk-table__cell">
-              <a class="govuk-link" href="{% url 'catalogue:reference_dataset_download' object.uuid 'csv' %}">
+              <a class="govuk-link" href="{% url 'datasets:reference_dataset_download' object.uuid 'csv' %}">
                 {{ object.slug }}-{{ object.published_version }}.csv
               </a>
             </td>
